Add raster layer raserization
Portions of raster layers can be rasterized based on the view of a
camera object.

+++ src/raster/position.rs
@@ -13,7 +13,7 @@ pub struct PixelPosition(pub (usize, usize));
 /// An position at which to draw something. Does not have to be inside the
 /// destination pixel collection.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct DrawPosition(pub (i32, i32));
+pub struct DrawPosition(pub (i64, i64));
 
 impl Add<PixelPosition> for PixelPosition {
     type Output = PixelPosition;
@@ -35,19 +35,19 @@ impl Add<DrawPosition> for PixelPosition {
     type Output = DrawPosition;
 
     fn add(self, rhs: DrawPosition) -> Self::Output {
-        DrawPosition((self.0 .0 as i32 + rhs.0 .0, self.0 .1 as i32 + rhs.0 .1))
+        DrawPosition((self.0 .0 as i64 + rhs.0 .0, self.0 .1 as i64 + rhs.0 .1))
     }
 }
-impl Add<(i32, i32)> for DrawPosition {
+impl Add<(i64, i64)> for DrawPosition {
     type Output = DrawPosition;
 
-    fn add(self, rhs: (i32, i32)) -> Self::Output {
+    fn add(self, rhs: (i64, i64)) -> Self::Output {
         DrawPosition((self.0 .0 + rhs.0, self.0 .1 + rhs.1))
     }
 }
 
-impl From<(i32, i32)> for DrawPosition {
-    fn from(p: (i32, i32)) -> Self {
+impl From<(i64, i64)> for DrawPosition {
+    fn from(p: (i64, i64)) -> Self {
         DrawPosition(p)
     }
 }
